Refactor: timestamp 컴포넌트화
