+++ src/components/chat/MySpeak.tsx
@@ -1,5 +1,6 @@
 import type { FC } from 'react';
 import { css } from '@emotion/react';
+import TimeStamp from './TimeStamp';
 
 interface MySpeakProps {
   content: string, date: number
@@ -11,7 +12,7 @@ const MySpeak: FC<MySpeakProps> = ({ content, date }) => (
       <span css={mySpeakCSS}>
         {content}
       </span>
-      <span css={timestampCSS}>{makeDate(date)}</span>
+      <TimeStamp timestamp={date} />
     </div>
   </span>
 );
@@ -27,22 +28,3 @@ const mySpeakCSS = css`
   border-radius: 10px 0px 10px 10px;
   padding: 12px;
 `;
-
-const timestampCSS = css`
-  color: #797C7B;
-  font-size: 10px;
-  padding: 3px;
-`;
-
-const makeDate = (date : number) => {
-  const dateObject = new Date(date);
-  let hour = dateObject.getHours();
-  const ampm = hour > 12 ? 'PM' : 'AM';
-  if (hour > 12) {
-    hour -= 12;
-  } else if (hour === 0) {
-    hour += 12;
-  }
-  const minute = dateObject.getMinutes();
-  return `${hour}:${minute} ${ampm}`;
-};
